+++ bin/cli.js
@@ -16,6 +16,9 @@ var argv = require('yargs')
   .alias("o", "output")
   .describe("output", "Change path to the generated tests.")
   .default("output", "tests")
+  .alias("r", "run-tests")
+  .describe("run-tests", "Run generated tests using `elm-test`.")
+  .default("run-tests", false)
   .describe("elm-test", "Path to elm-test.")
   .default("elm-test", path.join(__dirname, '../node_modules/.bin/elm-test'))
   .describe("elm-test-args", "Pass arguments to elm-test. f.e. `--elm-test-args=\"--report=junit\"`")
@@ -31,9 +34,14 @@ var model = runner.init(argv);
 
 runner.run(model, function(warnings) {
   warnings.map(runner.warnModule(model));
-  var status = runner.runElmTest(model);
-  if (status === 0) runner.cleanup(model);
-  runner.warnSummary(model, warnings);
-  process.exit(status);
+  if (model["run-tests"]) {
+    var status = runner.runElmTest(model);
+    if (status === 0) runner.cleanup(model);
+    runner.warnSummary(model, warnings);
+    process.exit(status);
+  } else {
+    runner.warnSummary(model, warnings);
+    process.exit(0);
+  }
 });